feat: refactor cell initialization, keyboard nav to a modifier
