+++ packages/components/src/components/hds/advanced-table/index.hbs
@@ -9,8 +9,7 @@
   role="grid"
   aria-describedby={{this._captionId}}
   {{style grid-template-columns=this.gridTemplateColumns}}
-  {{did-insert this.didInsert}}
-  {{will-destroy this.willDestroy}}
+  {{this._setUpObserver}}
 >
   <div id={{this._captionId}} class="sr-only hds-advanced-table__caption" aria-live="polite">
     {{@caption}}
