+++ packages/components/src/components/hds/advanced-table/index.ts
@@ -10,6 +10,7 @@ import { tracked } from '@glimmer/tracking';
 import type { ComponentLike } from '@glint/template';
 import { guidFor } from '@ember/object/internals';
 import config from 'ember-get-config';
+import { modifier } from 'ember-modifier';
 
 import {
   HdsAdvancedTableDensityValues,
@@ -282,7 +283,7 @@ export default class HdsAdvancedTable extends Component<HdsAdvancedTableSignatur
     return classes.join(' ');
   }
 
-  @action didInsert(element: HTMLDivElement): void {
+  private _setUpObserver = modifier((element: HTMLElement) => {
     const stickyGridHeader = element.querySelector(
       '.hds-advanced-table__thead.hds-advanced-table__thead--sticky'
     );
@@ -299,14 +300,13 @@ export default class HdsAdvancedTable extends Component<HdsAdvancedTableSignatur
 
       this._observer.observe(stickyGridHeader);
     }
-  }
 
-  @action willDestroy() {
-    super.willDestroy();
-    if (this._observer) {
-      this._observer.disconnect();
-    }
-  }
+    return () => {
+      if (this._observer) {
+        this._observer.disconnect();
+      }
+    };
+  });
 
   @action
   setSortBy(column: string): void {
