+++ packages/components/src/components/hds/advanced-table/td.ts
@@ -11,12 +11,7 @@ import { focusable, type FocusableElement } from 'tabbable';
 
 import type { HdsAdvancedTableHorizontalAlignment } from './types.ts';
 import { HdsAdvancedTableHorizontalAlignmentValues } from './types.ts';
-import {
-  didInsertGridCell,
-  handleGridCellKeyPress,
-  updateTabbableChildren,
-  onFocusTrapDeactivate,
-} from './helpers.ts';
+import { onFocusTrapDeactivate } from './helpers.ts';
 
 export const ALIGNMENTS: string[] = Object.values(
   HdsAdvancedTableHorizontalAlignmentValues
@@ -37,7 +32,6 @@ export interface HdsAdvancedTableTdSignature {
 export default class HdsAdvancedTableTd extends Component<HdsAdvancedTableTdSignature> {
   @tracked private _shouldTrapFocus = false;
   private _element!: HTMLDivElement;
-  private _observer: MutationObserver | undefined = undefined;
 
   // rowspan and colspan have to return 'auto' if not defined because otherwise the style modifier sets grid-area: undefined on the cell, which breaks the grid styles
   get rowspan(): string {
@@ -91,28 +85,7 @@ export default class HdsAdvancedTableTd extends Component<HdsAdvancedTableTdSign
     return cellFocusableElements[0];
   }
 
-  @action
-  didInsert(element: HTMLDivElement): void {
+  @action setElement(element: HTMLDivElement): void {
     this._element = element;
-    didInsertGridCell(element);
-    element.addEventListener('keydown', (event: KeyboardEvent) => {
-      handleGridCellKeyPress(event, this.enableFocusTrap);
-    });
-
-    this._observer = new MutationObserver(() => {
-      updateTabbableChildren(this._element, this._shouldTrapFocus);
-    });
-
-    this._observer.observe(this._element, {
-      childList: true,
-      subtree: true,
-    });
-  }
-
-  @action willDestroy() {
-    super.willDestroy();
-    if (this._observer) {
-      this._observer.disconnect();
-    }
   }
 }
