+++ packages/components/src/components/hds/advanced-table/th-sort.hbs
@@ -8,12 +8,14 @@
   role="columnheader"
   aria-rowspan={{@rowspan}}
   aria-colspan={{@colspan}}
-  {{did-insert this.didInsert}}
-  {{will-destroy this.willDestroy}}
+  {{hds-advanced-table-cell
+    handleEnableFocusTrap=this.enableFocusTrap
+    shouldTrapFocus=this._shouldTrapFocus
+    setCellElement=this.setElement
+  }}
   {{! @glint-expect-error - https://github.com/josemarluedke/ember-focus-trap/issues/86 }}
   {{focus-trap
     isActive=this._shouldTrapFocus
-    shouldSelfFocus=true
     focusTrapOptions=(hash
       onDeactivate=this.onFocusTrapDeactivate initialFocus=this.getInitialFocus clickOutsideDeactivates=true
     )
