+++ packages/components/src/components/hds/advanced-table/th-sort.ts
@@ -21,12 +21,7 @@ import type {
   HdsAdvancedTableThSortOrderLabels,
 } from './types.ts';
 import type { HdsAdvancedTableThButtonSortSignature } from './th-button-sort.ts';
-import {
-  didInsertGridCell,
-  handleGridCellKeyPress,
-  onFocusTrapDeactivate,
-  updateTabbableChildren,
-} from './helpers.ts';
+import { onFocusTrapDeactivate } from './helpers.ts';
 
 export const ALIGNMENTS: string[] = Object.values(
   HdsAdvancedTableHorizontalAlignmentValues
@@ -52,7 +47,6 @@ export default class HdsAdvancedTableThSort extends Component<HdsAdvancedTableTh
   private _labelId = guidFor(this);
   private _element!: HTMLDivElement;
   @tracked private _shouldTrapFocus = false;
-  private _observer: MutationObserver | undefined = undefined;
 
   get ariaSort(): HdsAdvancedTableThSortOrderLabels {
     switch (this.args.sortOrder) {
@@ -103,28 +97,7 @@ export default class HdsAdvancedTableThSort extends Component<HdsAdvancedTableTh
     return cellFocusableElements[0];
   }
 
-  @action
-  didInsert(element: HTMLDivElement): void {
+  @action setElement(element: HTMLDivElement): void {
     this._element = element;
-    didInsertGridCell(element);
-    element.addEventListener('keydown', (event: KeyboardEvent) => {
-      handleGridCellKeyPress(event, this.enableFocusTrap);
-    });
-
-    this._observer = new MutationObserver(() => {
-      updateTabbableChildren(this._element, this._shouldTrapFocus);
-    });
-
-    this._observer.observe(this._element, {
-      childList: true,
-      subtree: true,
-    });
-  }
-
-  @action willDestroy() {
-    super.willDestroy();
-    if (this._observer) {
-      this._observer.disconnect();
-    }
   }
 }
