+++ packages/components/src/components/hds/advanced-table/th.hbs
@@ -9,12 +9,14 @@
   aria-colspan={{@colspan}}
   aria-describedby={{@parentId}}
   {{style grid-row=this.rowspan grid-column=this.colspan padding-left=this.paddingLeft}}
-  {{did-insert this.didInsert}}
-  {{will-destroy this.willDestroy}}
+  {{hds-advanced-table-cell
+    handleEnableFocusTrap=this.enableFocusTrap
+    shouldTrapFocus=this._shouldTrapFocus
+    setCellElement=this.setElement
+  }}
   {{! @glint-expect-error - https://github.com/josemarluedke/ember-focus-trap/issues/86 }}
   {{focus-trap
     isActive=this._shouldTrapFocus
-    shouldSelfFocus=true
     focusTrapOptions=(hash
       onDeactivate=this.onFocusTrapDeactivate initialFocus=this.getInitialFocus clickOutsideDeactivates=true
     )
